+++ resources/lang/en/responses.php
@@ -15,6 +15,6 @@ return [
     'patch-invalid' => 'Unable to handle your request, there are non-existent fields in your request body.',
     'validation' => 'Validation error.',
     'authentication-required' => 'Authentication required, please try again with a Bearer.',
-    'maintenance' => 'Down for maintenance, we should be back very soon.',
+    'maintenance' => 'Down for maintenance, we should be back very soon, please check https://status.costs-to-expect.com for more information',
     'error' => 'Sorry, there has been an error, please try again later.'
 ];